// frontend/src/__generated__/models/RomioGameSchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RomioGameSchema = {
    id: string;
    title: string;
    system: string;
    systemName: string;
    rommSlug: string;
    browserCore?: (string | null);
    sourceCount: number;
    coverUrl?: (string | null);
    rating?: (number | null);
    ratingSource?: (string | null);
    awards: Array<string>;
    metadataSource?: (string | null);
};