// frontend/src/__generated__/models/RomioHomeSchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { RomioHomeSectionSchema } from './RomioHomeSectionSchema';
export type RomioHomeSchema = {
    sections: Array<RomioHomeSectionSchema>;
    language: 'en' | 'es' | 'fr' | 'ru' | 'zh' | 'ja';
};